marginalia_paratexta/requirements.txt
dj-database-url==0.5.0
Django==3.2.8
django-nested-admin==4.0.2
django-nested-inline==0.4.6
django-currentuser==0.5.3
django-polymorphic==3.1.0
django-threadlocals==0.10
django-treenode==0.20.1
django-filter==21.1
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gunicorn==20.1.0
django-jazzmin==2.6.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
requests==2.24.0
requests-oauthlib==1.3.0
wheel==0.38.1
whitenoise==5.3.0
matplotlib
plotly
